rf(context-test): Make context data structures available in schema/fixtures.test.ts
